refactor : 성재님 리뷰 반영
