--- src/main/java/com/opus/opus/modules/team/domain/dao/TeamSubmissionResult.java
@@ -1,10 +1,0 @@
-package com.opus.opus.modules.team.domain.dao;
-
-public record TeamSubmissionResult(
-        Long teamId,
-        String teamName,
-        String projectName,
-        String trackName,
-        Boolean isSubmitted
-) {
-}
